fix(tier4_autoware_utils): separate path_with_lane_id utils
